0.12.3dev: fix an unicode issue in the `[[Image]]` macro.
When given a project relative path (as in #10485), we shouldn't
produce an UTF-8 encoded `str` out of it, as this will trigger an
`UnicodeError` when mixed with `unicode` later on.

Modified all the Image macros to take some unicode input, to verify
the other code paths. Also removed the request parameters from the
image URL, in the `alt` and `title` attributes, to make them less
noisy.

Closes #10485.


git-svn-id: http://svn.edgewall.org/repos/trac/branches/0.12-stable@10911 af82e41b-90c4-0310-8c96-b1721e28e2e2