1.0.2dev: Don't add the datepicker when page is readonly. Refs #11069.
Previously, the datepicker would apppear when clicking on a `readonly` element (although this was entirely harmless since the form couldn't be saved). Also, the `readonly` attribute is explicitly assigned the `readonly` value since Genshi does not generate strict XHTML when `readonly = True` (genshi:#570). The value of the `disabled` attribute is handled explicitly in the code for consistency as well.

The permission variables have been extracted to the top of the template in preparation for fixing the fine-grained permission checks.


git-svn-id: http://svn.edgewall.org/repos/trac/branches/1.0-stable@12026 af82e41b-90c4-0310-8c96-b1721e28e2e2