1.0.2dev: Refactor admin panel permission checks and add functional tests. Refs #11069.
Remove redundant permission checks in `render_admin_panel` methods. The permissions are enforced through the return values of the `get_admin_panels` methods, and removing the redundant checks will make it easier to implement the necessary modifications for TracFineGrainedPermissions support.

For each of the admin panels an authorization test has been added to check that permissions are properly enforced.


git-svn-id: http://svn.edgewall.org/repos/trac/branches/1.0-stable@12027 af82e41b-90c4-0310-8c96-b1721e28e2e2